+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 - Fixed open screen in other dimensions than overworld
 - Fixed lever placing in dungeons
 - Fixed keep inventory option
+- Fixed water removing portal
 ### Changed:
 -